Extract admin handlers into custom hooks
Refactor AdminDashboard by extracting people and registration management handlers into dedicated custom hooks (useAdminPeopleActions and useAdminRegistrationActions). This reduces component complexity, improves code organization, and makes handlers more reusable and testable.
